Support saving kernel crash logs using boot2dump
Boot2dump is a new tiny operating system written by me, which is specialized
for saving a crash dump. Once Kerla panics, it jumps into the boot2dump's boot code.

Boot2dump reads an ext4 filesytem in a virtio-blk device, saves the dump into a file
and then it resets the computer.

It's an already working mechanism in my deployment in DigitalOcean!